[IMP] product_conversion: improve product conversion workflow
- Align fields in the Product Conversion wizard form view
- Update "Convert Into" domain to allow only storable (goods) products
- Enable product conversion from Inventory Valuation barcode flow

closes odoo/industry#2478

Task: 6054491
X-original-commit: 7dc50316e2d52974037e2ae77fc6ec8ddb77c4db
