+++ product_conversion/data/ir_model_fields.xml
@@ -184,7 +184,7 @@ for wizard in self:
     src_convert = src_ptavs.filtered(lambda p: p.attribute_id.x_is_convert_attribute)
     src_fixed = src_ptavs - src_convert
 
-    domain = [('id', '!=', source.id)]
+    domain = [('id', '!=', source.id), ('type', '=', 'consu')]
     if src_convert:
         domain.append(('product_tmpl_id', '=', source.product_tmpl_id.id))
         for product in wizard.env['product.product'].search(domain):
