+++ product_conversion/data/ir_ui_view.xml
@@ -101,12 +101,8 @@
                             'show_src_package': True
                         }"/>
                     <field name="x_quantity" string="Quantity"/>
-                </group>
-                <group>
                     <field name="x_keep_lot_sn" string="Keep Lot/SN" invisible="x_source_tracking not in ('serial', 'lot')"/>
                     <field name="x_new_lot_sn" string="New Lot/SN" invisible="x_source_tracking not in ('serial', 'lot') or x_keep_lot_sn" placeholder="No Lot/SN, product tracked by quantity."/>
-                </group>
-                <group>
                     <field name="x_product_convert_into" string="Convert into"
                         required="True"
                         domain="[('id', 'in', x_allowed_product_ids)]"
@@ -118,11 +114,10 @@
                             'list_view_ref': 'product_conversion.view_product_product_tree_convert',
                         }"
                     />
-                </group>
-                <group invisible="not x_show_destination_location">
                     <field name="x_destination_location_id" string="Destination Location"
                         domain="[('usage', '=', 'internal')]"
                         placeholder="Same as source location if empty."
+                        invisible="not x_show_destination_location"
                     />
                 </group>
                 <footer>
